requirements.txt
git+https://github.com/NAFTeam/NAFF.git@50d601a6763b87f29b7baa1927b389989aba4fc5
asyncpg==0.25.0
tortoise-orm==0.19.0
pytz==2022.1
websockets==10.3
python-dotenv==0.20.0
rapidfuzz==2.0.11
python-dateutil==2.8.2
flask==2.1.2
orjson==3.6.8
humanize==4.1.0
attrs==21.4.0
tomli==2.0.1
aioredis[hiredis]==2.0.1